% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-preprocessing.R
\name{add_indices_experiment_log}
\alias{add_indices_experiment_log}
\title{This hack assigns Index based on the number of occurances of hte same
Event and message
.e.g. the first recall start gets 1, seconds recall start gets 2}
\usage{
add_indices_experiment_log(df_test)
}
\description{
This hack assigns Index based on the number of occurances of hte same
Event and message
.e.g. the first recall start gets 1, seconds recall start gets 2
}
