% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-preprocessing.R
\name{convert_czech_to_en}
\alias{convert_czech_to_en}
\title{Converts czech names into english}
\usage{
convert_czech_to_en(items)
}
\arguments{
\item{items}{}
}
\value{

}
\description{
Converts czech names into english
}
