% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{filter_times.vremt}
\alias{filter_times.vremt}
\title{Filters the VREMT object for particular timestamps Keeps all the list
information, such as settings etc.}
\usage{
\method{filter_times}{vremt}(obj, times)
}
\arguments{
\item{obj}{vremt object}

\item{times}{times which to filter}
}
\value{
vremt object with only filtered times
}
\description{
Filters the VREMT object for particular timestamps Keeps all the list
information, such as settings etc.
}
