% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-visualisation.R
\name{geom_vremt_location}
\alias{geom_vremt_location}
\title{Plots locations and placement circles}
\usage{
geom_vremt_location(locations)
}
\arguments{
\item{locations}{character vector of locations to be plotted}
}
\value{
lsit with ggplot geoms
}
\description{
Plots locations and placement circles
}
