% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-analysis.R
\name{get_arm_distances}
\alias{get_arm_distances}
\title{Return distances to all arms}
\usage{
get_arm_distances(location)
}
\arguments{
\item{location}{numeric(3) with position}
}
\value{
vector 5 with distances to each arm
}
\description{
Return distances to all arms
}
