% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-analysis.R
\name{get_closest_arm}
\alias{get_closest_arm}
\title{Returns which arm is closest}
\usage{
get_closest_arm(location)
}
\arguments{
\item{location}{Location is set as vector 3}
}
\value{

}
\description{
this is still better than get last get_last_bridge_entered
due to complications with that function. See its manual for better description
}
