% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_collected_items}
\alias{get_collected_items}
\title{Returns list of collected items from given object action log}
\usage{
get_collected_items(phase_obj)
}
\arguments{
\item{phase_obj}{Generally preselected vremt object. If you pass the entire
object, not just particular phase, all items will be returned.}
}
\value{

}
\description{
Returns list of collected items from given object action log
}
