% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_last_bridge_entered}
\alias{get_last_bridge_entered}
\title{Returns which was the last arm entered based on bridge crossing}
\usage{
get_last_bridge_entered(obj, time)
}
\arguments{
\item{obj}{vremt object. Usually specific recallPlacement or pickup phase}

\item{time}{time of interest}
}
\description{
this is prone to error due to the fact how the triggers are
recorded. Participant can technically enter a bridge and then exit it withou
ever stepping into the specific arm - we could solve for this by tracking
the same enter exit trigger twice, but this does not solve the issue at hand
as it would be
}
\examples{
I am getting both enter and exit so that there is less
errors due to phase switching (e.g. participant dropping an ittem
before any movement would potentially trigger last trigger)
}
