% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_phase_data}
\alias{get_phase_data}
\title{Returns portion of the object relevant to thegiven phase}
\usage{
get_phase_data(obj, phase, index = NA)
}
\arguments{
\item{phase}{Name of the phase as appears in the experiment log in the
Message parameter (e.g. exploration, recall, recallPlacement). Some phases
(recall) have multiple "subphases"}

\item{index}{Index of the phase (one based) in case there are multiple phases
of the same name}
}
\value{

}
\description{
Returns portion of the object relevant to thegiven phase
}
