% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_phase_task_index}
\alias{get_phase_task_index}
\title{Returns the recall task index in the given object}
\usage{
get_phase_task_index(phase_obj, zero_based = TRUE)
}
\arguments{
\item{phase_obj}{filtered phase vremt object. Use get_phase_}
}
\value{
ONE BASED integer. This
}
\description{
The recallPhases are randomized, with the specific task being
encoded in the particular pohase's action log. This extract what settings
ran during this particular phase. Only works on recall phase properly.
Exploration and pickup phases come in Task 0 1 2 ordcer always. This returns
ONE based task order for less interference with R subsetting.
}
