% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_phase_task_settings}
\alias{get_phase_task_settings}
\title{Returns the settings for the given phase object}
\usage{
get_phase_task_settings(phase_obj)
}
\arguments{
\item{phase_obj}{}
}
\value{

}
\description{
wrapper around get_task_settings and get_phase_task_index.
}
