% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_phase_time}
\alias{get_phase_time}
\title{Returns times of given phase.}
\usage{
get_phase_time(obj, phase, index = NA)
}
\arguments{
\item{obj}{vremt object}

\item{phase}{name of the phase as it stands in the experiment_log. e.g. "recall"}

\item{order}{index of hte phase in case there were multiple phases
of the same number}
}
\value{
returns numeric(2) of time start
}
\description{
Returns times of given phase.
}
