% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_task_settings}
\alias{get_task_settings}
\title{Returns the task settings for the object. If index is passed, returns only
settings for a particular trial}
\usage{
get_task_settings(obj, index = NA)
}
\arguments{
\item{obj}{vremt object}

\item{index}{numeric designation of the task order (usually between 1-3)}
}
\value{
list with items and locations fields
}
\description{
Returns the task settings for the object. If index is passed, returns only
settings for a particular trial
}
