% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-data.R
\docType{data}
\name{ITEM_CODES}
\alias{ITEM_CODES}
\title{Dataset with Czech and English codes for items.}
\format{
A data frame with 25 values and 4 variables:
\describe{
  \item{name_cz}{Czech item name}
  \item{name_en}{English item name}
  \item{position}{Position of the item given as a string (X, Y, Z)}
  \item{details}{some specific description of the item}
  ...
}
}
\usage{
ITEM_CODES
}
\description{
This dataset allows translation of Czech codes from logs to more error
error free English variants. It is primarily used during data loading
}
\keyword{datasets}
