% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-loading.R
\name{load_vremt_experiments}
\alias{load_vremt_experiments}
\title{Loads all vremt experiments from given folder}
\usage{
load_vremt_experiments(
  folder,
  version = 2024,
  preprocess = TRUE,
  flatten = FALSE
)
}
\arguments{
\item{preprocess}{should the loaded files be preprocessed}

\item{flatten}{in case a single experiment is found, should the results be
flattened (list[[1]] becomes the result)}

\item{folders}{Which folder to load}
}
\value{
list of class vremt
}
\description{
Loads all vremt experiments from given folder
}
