% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-plotting.R
\name{plot_path.vremt}
\alias{plot_path.vremt}
\title{Plots path}
\usage{
\method{plot_path}{vremt}(obj, ...)
}
\arguments{
\item{obj}{vremt object}

\item{...}{ggplot arguments}
}
\value{

}
\description{
Wrapper around \code{\link{plot_path}{navr}}
}
