% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-analysis.R
\name{vremt_collection_performance}
\alias{vremt_collection_performance}
\title{Analyses the recall item collection performance}
\usage{
vremt_collection_performance(recallItems)
}
\arguments{
\item{recallItems}{vremt recallItem object receoved with
\code{\link{get_recallItems_data}}}
}
\value{
list with resutls
}
\description{
Analyses the recall item collection performance
}
