% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-analysis.R
\name{vremt_placement_performance}
\alias{vremt_placement_performance}
\title{Analyses the recall placement phase}
\usage{
vremt_placement_performance(recallPlacement)
}
\arguments{
\item{recallPlacement}{vremt placement object got with
\code{\link{get_recallPlacement_data}}}

\item{index}{order of the phase to analyse}
}
\value{
list with results
}
\description{
Analyses the recall placement phase
}
