% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-preprocessing.R
\name{decompose_task}
\alias{decompose_task}
\title{Decomposes task as listed in the experiment info into data.frame}
\usage{
decompose_task(task)
}
\arguments{
\item{task}{character with task settings}
}
\value{
data.frame (island, locations, )
}
\description{
Decomposes task as listed in the experiment info into data.frame
}
