% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-getters.R
\name{get_location_position}
\alias{get_location_position}
\alias{get_item_position}
\alias{get_arm_position}
\title{Returns the position of a particular location (or item)}
\usage{
get_location_position(locations, simplify = TRUE)

get_item_position(items, simplify = TRUE)

get_arm_position(arms = seq_len(5), simplify = TRUE)
}
\arguments{
\item{locations}{name of the location defining the location}

\item{simplify}{only valid in case of a single location return. If true,
only a vector is returned. Otherwise even single location is returned as a
data.frame}

\item{arms}{number 1-5 defining the arm of the island. Defaults to all arms}

\item{item}{name of the item}
}
\value{
named numeric vector or NULL if no location of such name is found or
a data.frame if
}
\description{
either location or item has to be passed. The returned location
is the same for both location name and item name (e.g. both cemetery as
a location and a globe as an item have the same position)
}
\section{Functions}{
\itemize{
\item \code{get_item_position}: getting item position

\item \code{get_arm_position}: returns position of a given arm
}}

\examples{
get_location_position("cemetery")
}
