% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-data.R
\docType{data}
\name{LOCATION_ITEM}
\alias{LOCATION_ITEM}
\title{Correct answers binding specific locations and items to a concrete arm}
\format{
A data frame with 15 values and 3 variables:
\describe{
  \item{arm}{Numbering order of the arm with the location (1-5)}
  \item{location}{English name of the location}
  \item{item}{Name of the item in EN code}
  ...
}
}
\usage{
LOCATION_ITEM
}
\description{
This dataset allows analysis of correct answers as the items are effectively
coded as locations not itmes (each item always belong to a single location)
}
\keyword{datasets}
