% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-preprocessing.R
\name{preprocess_vremt}
\alias{preprocess_vremt}
\title{Preprocesses vremt object}
\usage{
preprocess_vremt(obj)
}
\arguments{
\item{obj}{object of class vremt. loaded with load_vremt_experiments}
}
\value{

}
\description{
Preprocesses vremt object
}
