% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vremt-analysis.R
\name{vremt_placement_performance}
\alias{vremt_placement_performance}
\title{Analyses the recall placemnent phase}
\usage{
vremt_placement_performance(obj, index)
}
\arguments{
\item{obj}{vremt object}

\item{index}{order of the phase to analyse}
}
\value{
list with results
}
\description{
Analyses the recall placemnent phase
}
